Exception handling - delete

+++ src/main/java/com/gustavostolze/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.gustavostolze.course.entities.User;
 import com.gustavostolze.course.repositories.UserRepository;
+import com.gustavostolze.course.services.exceptions.DataBaseException;
 import com.gustavostolze.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -15,24 +17,33 @@ public class UserService {
 
 	@Autowired
 	private UserRepository userRepository;
-	
+
 	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		if (userRepository.existsById(id)) {
+			try {
+				userRepository.deleteById(id);
+			} catch (DataIntegrityViolationException e) {
+				throw new DataBaseException(e.getMessage());
+			}
+		}
+		else {
+			throw new ResourceNotFoundException(id);
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = userRepository.getReferenceById(id);
 		updateData(entity, obj);
